// apps/yearn/src/components/shared/types/mixed.ts
export type TDict<T> = { [key: string]: T }
export type Dict<T> = TDict<T>
export type TNDict<T> = { [key: number]: T }
// ...
export type TNormalizedBN = { raw: bigint; normalized: number; display: string; decimals: number }

/*******************************************************************************
 ** A classic ERC20 token & the one wrapped by chainID
 ******************************************************************************/
export type TToken = {
  address: TAddress
  name: string
  symbol: string
  decimals: number
  chainID: number
  logoURI?: string
  value: number
  balance: TNormalizedBN
}
export type TChainTokens = TNDict<TDict<TToken>>